move core logic of convert_int_var_order_encoding to order::
